Rename html_reporter and json_processor
